requirements.txt
nameko==3.0.0-rc11
speedtest-cli
loguru
pytest
pytest-cov